Add flask server, roadmap generation, authentication, and UI improvements
- Implemented LLM-based roadmap generation in the backend.
- Added retype password validation in the signup form.
- Improved modal behavior by clearing messages on close or mode switch.
- Added error handling and validation for LLM responses.
